requirements.txt
h5py==2.10.0
html5lib==1.0.1
opencv-python==4.2.0.32
Pillow==6.2.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pydeck==0.3.0
pydocstyle==3.0.0
pyflakes==2.1.1
Pygments==2.5.2
streamlit==0.55.2
sympy==1.5.1
tensorboard==2.1.0
tensorflow==2.0.0b0
tensorflow-estimator==2.0.0
tensorflow-hub==0.8.0
tqdm==4.44.1
